requirements.txt
asgiref==3.5.2
autopep8==1.6.0
backports.zoneinfo==0.2.1
Django==4.0.5
django-phonenumber-field==6.3.0
phonenumbers==8.12.53
pycodestyle==2.8.0
sqlparse==0.4.2
toml==0.10.2
